updateGoods correct

+++ src/service/GoodsTypeService.ts
@@ -44,9 +44,13 @@ export class GoodsTypeService {
     }
 
     async deleteGoodsType(id:number):Promise<void>{
-        let goodsType: GoodsType = await this.goodsTypeRepository.findOneById(id)
+        let goodsType: GoodsType = await this.goodsTypeRepository.findOneById(id,{relations:['properties','goodses']})
         if (!goodsType) {
             throw new HttpException('指定id=' + id + '商品类型不存在', 404)
+        }
+        /* 商品类型下还有商品则不能删除，必须先解除两者关系 */
+        if(goodsType.goodses&&goodsType.goodses.length>0){
+            throw new HttpException('指定商品类型下存在商品，请先将商品移动到其他商品类型下再删除',404)
         }
         try {
             await this.goodsTypeRepository.remove(goodsType)
